[MOD]backup: optional scp
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@25228 b456876b-0849-0410-b77d-98878d47e9d5

+++ doc/devtools/backupdb.sh
@@ -26,11 +26,12 @@ DBUSER=${user_tiki:-'root'}
 DBPASS=${pass_tiki:-''}
 
 NOWDATE=`date +%Y-%m-%d`
-DUMPFILE="site_backup.$VIRTUAL.$DATE.sql"
+DUMPFILE="$DBNAME-$NOWDATE-sql.bz2"
 
 cd $ARCHIVEDIR
 
 mysqldump -Q -u $DBUSER -h$DBHOST -p$DBPASS $DBNAME | bzip2 -c > $DUMPFILE
-scp -C $DUMPFILE $SCPTO
-rm $DUMPFILE
-
+if [ "$SCPTO" ]; then 
+	scp -C $DUMPFILE $SCPTO
+	rm $DUMPFILE
+fi
